10.7.5.3: Using wrapper type for Session to avoid messing
Refer to original doc: Session is powerful but, taken as is, it is a brittle foundation to build your application state- handling on. We are accessing data using a string-based API, being careful to use the same keys and types on both sides - insertion and retrieval.

It works when the state is very simple, but it quickly degrades into a mess if you have several routes accessing the same data - how can you be sure that you updated all of them when you want to evolve the schema? How do we prevent a key typo from causing a production outage?

Tests can help, but we can use the type system to make the problem go away entirely. We will build a strongly-typed API on top of Session to access and modify the state - no more string keys and type casting in our request handlers.